requirements.txt
flask
python-dotenv
# Core LangChain (v0.3 compatible)
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-huggingface==0.1.0
langchain-chroma==0.1.4
# Database & Tools
chromadb
pypdf
sentence-transformers
deep-translator
pyttsx3
